skip permission, not found errors on propfind
We used to skip only certain directory errors when running propfind.
Instead, skip a class of errors for all files, directory or not:
 - Permission defined
 - Not found
 - Path error

All of these errors indicate that the underlying filesystem is unable to
give the server access to the desired file for whatever reason. This
could be because it doesn't have permission, or the file was removed, or
something else we aren't aware of. Fundamentally, the file system API
can only abstract so much, so it's best to be robust when handling
errors in a propfind. Otherwise, one skunky file can break an entire
directly, and even a whole share if it's at the root.